refactor: Restructure widget modules by creating controls, inputs, and scrollable files
